+++ docs/src/plugin-vue/changelog.md
@@ -12,6 +12,16 @@ outline: 2
 
 _Nothing yet._
 
+## 0.2.0
+
+### Breaking Changes
+
+- Rename `StoreRefOptions.writeDefaults` to [`StoreRefOptions.writeDefault`](https://tb.dev.br/tauri-store/js-docs/plugin-vue/interfaces/StoreRefOptions.html#writedefault).
+
+### Bug fixes
+
+- Auto start global store.
+
 ## 0.1.1
 
 ### Features